==> 01_Codes/Revanth/02_Messidor_Processing.ipynb <==
To read the zipped image files and convert them all to a standard size.
once converted, they are stored in the relevant folder.

# importing all the required libraries
import numpy as np
import pandas as pd
import zipfile
import os
import shutil
from PIL import Image
import pickle

### Configurations
modify the below configurations to suit your needs, beyond this, the rest of the code should work as is.

# setting default configuration options
pd.set_option("mode.copy_on_write", True)
desired_size = (512, 512)  # Modify this if required

Specify the Path to the folders as per your system in case running this file

# Repo Paths, No need to change these as they are present in the repo and are relative paths
path_combine = "../02_Data/Extra/messidor_combine.parquet"  # path to the combined excel DF (present in repo)
path_mapping = "../02_Data/01_messidor_mapping.parquet"  # path to the mapping parquet file (present in repo)

# Absolute Paths, Change these to the paths in your system
path_source = "/Users/revanth/Downloads/Messidor/"  # path to the source folder (Regular Folder not Zip)
path_dest = "/Users/revanth/Documents/Messidor_Data/"  # path to the destination folder (Regular Folder not Zip)

No need to change the below code beyond this point

## Loading Mapping File

# Loading the Combined Mapping File
labels = pd.read_parquet(path_combine)
labels.sample(5)

# adding empty column to store the original image size
labels["Original_Size"] = None

## Identifying Files

# lsiting all the files in the path_source folder
files = os.listdir(path_source)
files

# filtering for only zip files
zip_files = [
    file for file in files if file.startswith("Base") and file.endswith(".zip")
]
zip_files.sort()
zip_files

# we should have 12 zip files
assert len(zip_files) == 12

### Extracting the zip files

# creating a temporary directory to extract the files

# delete the temp directory if it already exists
if os.path.exists(path_source + "temp_dir/"):
    os.system("rm -rf " + path_source + "temp_dir/")
os.mkdir(path_source + "temp_dir/")
temp_dir = path_source + "temp_dir/"

# extracting the contents of the zip files to the temporary directory
for file in zip_files:
    with zipfile.ZipFile(path_source + file, "r") as zip_ref:
        zip_ref.extractall(temp_dir)
        print(f"Files from {file} extracted successfully")

# if temp_dir has sub-folders, extract the contents of the subfolders to the temp_dir

# traverse root directory, and list directories as dirs and files as files
for root, dirs, files in os.walk(temp_dir):
    path = root.split(os.sep)

    for file in files:
        if not os.path.isdir(file):

            # move file from nested folder into the base folder
            shutil.move(os.path.join(root, file), os.path.join(temp_dir, file))

### cleanup

for root, dirs, files in os.walk(temp_dir):
    # Deleting the sub-folders
    for dir in dirs:
        shutil.rmtree(os.path.join(root, dir))

    # deleting all the excel files and DS_Store files
    for file in files:
        if file.endswith(".xls"):
            os.remove(os.path.join(root, file))
        if file == ".DS_Store":
            os.remove(os.path.join(root, file))

# check if all the files are loaded with the correct names
num_base = 0
unique_start = set()
num_non_base = 0

for name in os.listdir(temp_dir):
    if "Base" in name:
        num_base += 1
        unique_start.add(name.split("/")[0])
    else:
        num_non_base += 1
    if name not in labels["Image_ID"].values:
        print(f"Image {name} not found in the labels")

print(f"Number of Base Images: {num_base}")
print(f"Number of Non-Base Images: {num_non_base}")

# we should have 1200 non-base images and 0 base images
assert num_base == 0
assert num_non_base == 1200

# view images which are loaded with the incorrect names
unique_start

## Image Processing

# check if the target folder exists, if not create it
if not os.path.exists(path_dest):
    os.mkdir(path_dest)
    print("Destination folder created successfully")
else:
    print("Destination folder already exists")

# Deleting the exisiting files in the destination folder if any
if os.path.exists(path_dest + "Resized/"):
    os.system("rm -rf " + path_dest + "Resized/")
    print("Existing Resized folder deleted successfully")
os.mkdir(path_dest + "Resized/")


# Deleting the exisiting files in the destination folder if any
if os.path.exists(path_dest + "Raw/"):
    os.system("rm -rf " + path_dest + "Raw/")
    print("Existing Raw folder deleted successfully")
os.mkdir(path_dest + "Raw/")

# creating sub-folders for Train and Test
os.mkdir(path_dest + "Resized/" + "Train/")
os.mkdir(path_dest + "Resized/" + "Test/")
print("Train and Test folders created successfully in Resized folder")

# creating sub-folders for Train and Test in Raw folder
os.mkdir(path_dest + "Raw/" + "Train/")
os.mkdir(path_dest + "Raw/" + "Test/")
print("Train and Test folders created successfully in Raw folder")

path_train_resized = path_dest + "Resized/" + "Train/"
path_test_resized = path_dest + "Resized/" + "Test/"
path_train_raw = path_dest + "Raw/" + "Train/"
path_test_raw = path_dest + "Raw/" + "Test/"

# go through the temp_dir and process the images and save them in the target directory
nums = 0  # to keep track of the number of images processed
err_num = 0  # to keep track of the number of images that caused an error

path = temp_dir
for files in os.listdir(path):
    if files.endswith(".tif"):
        img = Image.open(path + files)
        img_split = None  # so store if the image is test or train
        try:
            # add original size to the labels dataframe
            idx = labels[labels["Image_ID"] == files].index[0]
            labels.at[idx, "Original_Size"] = img.size

            # if the image is not marked as iclude, skip it
            if labels.at[idx, "Include"] == 0:
                continue
            # get the split of the image
            img_split = labels.at[idx, "Split"]

            # save the raw image in the respective folder
            if img_split == "Train":
                img.save(path_train_raw + files)
            elif img_split == "Test":
                img.save(path_test_raw + files)

            # resize the image
            img = img.resize(desired_size)

            # save the image
            if img_split == "Train":
                img.save(path_train_resized + files)
            elif img_split == "Test":
                img.save(path_test_resized + files)

            nums += 1

        except:
            print(f"Error processing {files}")
            err_num += 1
            if err_num > 10:
                # breaking so that we can fix the error before trying more images
                print("Too many errors!, breaking")
                break
            continue
    if nums % 100 == 0:
        print(f"processed {nums} images")
print(f"Processed {nums} images with {err_num} errors")

## Sanity Checks

# We should have 1200 images
assert nums == labels["Include"].sum()

# We should have 0 errors
assert err_num == 0

# Train folder should have same number of images as the Train split
assert len(os.listdir(path_train_resized)) == len(
    labels[(labels["Split"] == "Train") & labels["Include"]]
)
assert len(os.listdir(path_train_raw)) == len(
    labels[(labels["Split"] == "Train") & labels["Include"]]
)

# Test folder should have same number of images as the Test split
assert len(os.listdir(path_test_resized)) == len(
    labels[(labels["Split"] == "Test") & labels["Include"]]
)
assert len(os.listdir(path_test_raw)) == len(
    labels[(labels["Split"] == "Test") & labels["Include"]]
)

# Saving and Cleanup

# viewing the labels dataframe
labels.sample(5)

# saving the labels dataframe as a parquet file
labels.to_parquet(path_mapping, index=False)

# saving labels df as excel and parquet in both sub-folders
labels.to_parquet(path_dest + "/Raw/" + "messidor_mapping.parquet", index=False)
labels.to_excel(path_dest + "/Raw/" + "messidor_mapping.xlsx", index=False)
labels.to_parquet(path_dest + "/Resized/" + "messidor_mapping.parquet", index=False)
labels.to_excel(path_dest + "/Resized/" + "messidor_mapping.xlsx", index=False)

# cleaning up the temporary directory
shutil.rmtree(temp_dir)
