==> 01_Codes/Revanth/06_Basic_Models.ipynb <==
# importing required libraries
# Sklearn
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, confusion_matrix, classification_report

import matplotlib.pyplot as plt
import numpy as np
import time
import warnings
import pandas as pd
import os
import parquet
from PIL import Image

# Tensor Flow
# tensorflow
import tensorflow as tf
from tensorflow.keras.applications.resnet50 import preprocess_input

tf.config.list_physical_devices("CPU")

# Setting default options
# setting default configuration options
pd.set_option("mode.copy_on_write", True)
warnings.simplefilter(action="ignore")  # , category=FutureWarning)

## Loading Data

path = "../aws_s3/Messidor_Resized/"

test_files = os.listdir(path + "Test")
print(test_files[:5])

train_files = os.listdir(path + "Train")
print(train_files[:5])

mapping = pd.read_parquet("../02_Data/01_messidor_mapping.parquet")
mapping.sample(5)

X_train = np.array(
    [
        (np.array(Image.open(path + "Train/" + i))[None, ...] / 255.0).flatten()
        for i in train_files
    ]
)
y_train = np.array(
    [
        mapping.loc[mapping["Image_ID"] == i, "Retinopathy_Grade"].values[0]
        for i in train_files
    ]
)

X_test = np.array(
    [
        (np.array(Image.open(path + "Test/" + i))[None, ...] / 255.0).flatten()
        for i in test_files
    ]
)
y_test = np.array(
    [
        mapping.loc[mapping["Image_ID"] == i, "Retinopathy_Grade"].values[0]
        for i in test_files
    ]
)

X_train.shape

y_train.shape

### Random Forest

# Initializing the Model
model_rf_raw = RandomForestClassifier(n_estimators=100)

# Fitting the Model
t = time.time()
model_rf_raw.fit(X_train, y_train)
print(f"Training Time: {time.time() - t:.2f} Seconds")

# Predicting the Test Data
t = time.time()
y_pred = model_rf_raw.predict(X_test)
print(f"Prediction Time: {time.time() - t:0.2f} Seconds")

# Model Evaluation
print(f"Accuracy: {accuracy_score(y_test, y_pred):.2%}")
print(f"Confusion Matrix: \n{confusion_matrix(y_test, y_pred)}")

### Logistic Regression

# Initializing the Model
model_lr_raw = LogisticRegression(
    max_iter=500, multi_class="multinomial", solver="lbfgs", random_state=42
)

# Fitting the Model
t = time.time()
model_lr_raw.fit(X_train, y_train)
print(f"Training Time: {time.time() - t:0.2f} Seconds")

t = time.time()
y_pred = model_lr_raw.predict(X_test)
print(f"Prediction Time: {time.time() - t:0.2f} Seconds")

# Model Evaluation
print(f"Accuracy: {accuracy_score(y_test, y_pred):.2%}")
print(f"Confusion Matrix: \n{confusion_matrix(y_test, y_pred)}")

### RESNET

# Load a pretrained model
model = tf.keras.applications.resnet50.ResNet50(
    include_top=False,  # take any input shape, not just 224x224
    weights="imagenet",  # use the imagenet pretrained model
    pooling="avg",  # use global average pooling to get a 2048 vector from our image
)

%%capture
def preprocess(img):
    processed_img = preprocess_input(  # ResNet has its own preprocesser
        tf.image.resize(  # Images need to be in (224, 224) dimension
            np.array(img), (224, 224)  # Turn PIL image object into array
        )
    )[None]
    return processed_img


# use resnet to get the embeddings
X_train_resnet = np.array(
    [model.predict(preprocess(Image.open(path + "Train/" + i))) for i in train_files]
).reshape(-1, 2048)
X_test_resnet = np.array(
    [model.predict(preprocess(Image.open(path + "Test/" + i))) for i in test_files]
).reshape(-1, 2048)

### Random Forest - ResNet Optimized

# Initializing the Model
model_rf_resnet = RandomForestClassifier(n_estimators=100)

# Fitting the Model
t = time.time()
model_rf_resnet.fit(X_train_resnet, y_train)
print(f"Training Time: {time.time() - t:.2f} Seconds")

# predicting the test data
t = time.time()
y_pred = model_rf_resnet.predict(X_test_resnet)
print(f"Prediction Time: {time.time() - t:0.2f} Seconds")

# Model Evaluation
print(f"Accuracy: {accuracy_score(y_test, y_pred):.2%}")
print(f"Confusion Matrix: \n{confusion_matrix(y_test, y_pred)}")

### Logistic Regression - ResNet Optimized

# Initializing the Model
model_lr_resnet = LogisticRegression(
    max_iter=200, multi_class="multinomial", solver="lbfgs", random_state=42
)

# Fitting the Model
t = time.time()
model_lr_resnet.fit(X_train_resnet, y_train)
print(f"Training Time: {time.time() - t:0.2f} Seconds")

# predicting the test data
t = time.time()
y_pred = model_lr_resnet.predict(X_test_resnet)
print(f"Prediction Time: {time.time() - t:0.2f} Seconds")

# Model Evaluation
print(f"Accuracy: {accuracy_score(y_test, y_pred):.2%}")
print(f"Confusion Matrix: \n{confusion_matrix(y_test, y_pred)}")
